+++ alacritty/src/config/bindings.rs
@@ -161,6 +161,10 @@ pub enum Action {
     /// Hide the Alacritty window.
     Hide,
 
+    /// Hide all windows other than Alacritty on macOS.
+    #[cfg(target_os = "macos")]
+    HideOtherApplications,
+
     /// Minimize the Alacritty window.
     Minimize,
 
@@ -685,6 +689,7 @@ pub fn platform_key_bindings() -> Vec<KeyBinding> {
         C, ModifiersState::LOGO; Action::Copy;
         C, ModifiersState::LOGO, +BindingMode::VI, ~BindingMode::SEARCH; Action::ClearSelection;
         H, ModifiersState::LOGO; Action::Hide;
+        H, ModifiersState::LOGO | ModifiersState::ALT; Action::HideOtherApplications;
         M, ModifiersState::LOGO; Action::Minimize;
         Q, ModifiersState::LOGO; Action::Quit;
         W, ModifiersState::LOGO; Action::Quit;
@@ -1157,7 +1162,7 @@ impl<'a> de::Deserialize<'a> for ModsWrapper {
             type Value = ModsWrapper;
 
             fn expecting(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-                f.write_str("a subset of Shift|Control|Super|Command|Alt|Option")
+                f.write_str("None or a subset of Shift|Control|Super|Command|Alt|Option")
             }
 
             fn visit_str<E>(self, value: &str) -> Result<ModsWrapper, E>